Updated docs after refactoring
A lot of the cypher internals are being shuffled around, this commit
updates the cypher docs to cope with that.
