+++ cypher/cypher-docs/src/test/scala/org/neo4j/cypher/docgen/tooling/RestartableDatabase.scala
@@ -21,9 +21,8 @@ package org.neo4j.cypher.docgen.tooling
 
 import org.neo4j.cypher.ExecutionEngineHelper
 import org.neo4j.cypher.docgen.ExecutionEngineFactory
-import org.neo4j.cypher.internal.ExecutionEngine
-import org.neo4j.cypher.internal.compatibility.v3_3.runtime.executionplan.InternalExecutionResult
 import org.neo4j.cypher.internal.helpers.GraphIcing
+import org.neo4j.cypher.internal.{ExecutionEngine, InternalExecutionResult}
 import org.neo4j.cypher.javacompat.internal.GraphDatabaseCypherService
 import org.neo4j.test.{TestEnterpriseGraphDatabaseFactory, TestGraphDatabaseFactory}
