+++ cypher/graphgist/src/main/scala/org/neo4j/cypher/internal/DocsExecutionEngine.scala
@@ -22,7 +22,6 @@ package org.neo4j.cypher.internal
 import java.util.{Map => JavaMap}
 
 import org.neo4j.cypher.SyntaxException
-import org.neo4j.cypher.internal.compatibility.v3_3.runtime.executionplan.InternalExecutionResult
 import org.neo4j.kernel.GraphDatabaseQueryService
 import org.neo4j.kernel.impl.query.TransactionalContext
 import org.neo4j.logging.{LogProvider, NullLogProvider}
