+++ cypher/refcard-tests/src/test/scala/org/neo4j/cypher/docgen/refcard/ExampleTest.scala
@@ -22,7 +22,7 @@ package org.neo4j.cypher.docgen.refcard
 import org.junit.Ignore
 import org.neo4j.cypher.QueryStatisticsTestSupport
 import org.neo4j.cypher.docgen.RefcardTest
-import org.neo4j.cypher.internal.compatibility.v3_3.runtime.executionplan.InternalExecutionResult
+import org.neo4j.cypher.internal.InternalExecutionResult
 
 @Ignore
 class ExamplesTest extends RefcardTest with QueryStatisticsTestSupport {
